fixes the developer and changing in the populate file
