+++ arability/app/models/developer.rb
@@ -3,7 +3,7 @@ class Developer < ActiveRecord::Base
   belongs_to :gamer 
   has_one :my_subscription
   has_many :api_keys
-
+  has_and_belongs_to_many :keywords, uniq: true
   has_many :shared_projects
   has_many :projects_shared, :through => :shared_projects, :source => "project"
   has_many :projects, :foreign_key => "owner_id"
