+++ arability/lib/tasks/populate.rake
@@ -82,28 +82,28 @@ namespace :db do
     l=Synonym.create(name: "شارك", keyword_id: share.id, approved: true, is_formal: true)
     m=Synonym.create(name: "اشترك", keyword_id: share.id, approved: true, is_formal: true)
     n=Synonym.create(name: "بووست", keyword_id: post.id, approved: true, is_formal: false)
-    o=Synonym.create(name: "يلكز", keyword_id: poke.id, approved: true, is_formal: true)
+    o=Synonym.create(name: "يلكز", keyword_id: poke.id, approved: true, is_formal: nil)
     p=Synonym.create(name: "تابع", keyword_id: subscribe.id, approved: true, is_formal: true)
     q=Synonym.create(name: "اشترك", keyword_id: subscribe.id, approved: true, is_formal: true)
     w=Synonym.create(name: "انتشروا", keyword_id: broadcast.id, approved: true, is_formal: true)
-    x=Synonym.create(name: "كومبيل", keyword_id: compile.id, approved: true, is_formal: false)
+    x=Synonym.create(name: "كومبيل", keyword_id: compile.id, approved: true, is_formal: nil)
     y=Synonym.create(name: "يعجبني", keyword_id: like.id, approved: true, is_formal: true)
-    z=Synonym.create(name: "احبه", keyword_id: like.id, approved: true, is_formal: true)
+    z=Synonym.create(name: "احبه", keyword_id: like.id, approved: true, is_formal: nil)
 
     a1=Synonym.create(name: "ابحث", keyword_id: find.id, approved: true, is_formal: true)
     a2=Synonym.create(name: "دور", keyword_id: find.id, approved: true, is_formal: false)
     a3=Synonym.create(name: "دورلي", keyword_id: search.id, approved: true, is_formal: false)
-    a4=Synonym.create(name: "سيرش", keyword_id: search.id, approved: true, is_formal: false)
-    a5=Synonym.create(name: "انشر", keyword_id: onshor.id, approved: true, is_formal: true)
+    a4=Synonym.create(name: "سيرش", keyword_id: search.id, approved: true, is_formal: nil)
+    a5=Synonym.create(name: "انشر", keyword_id: onshor.id, approved: true, is_formal: nil)
     a6=Synonym.create(name: "اضغط", keyword_id: doss.id, approved: true, is_formal: true)
-    a7=Synonym.create(name: "احب ده", keyword_id: yagebony.id, approved: true, is_formal: false)
+    a7=Synonym.create(name: "احب ده", keyword_id: yagebony.id, approved: true, is_formal: nil)
     a8=Synonym.create(name: "مركبة", keyword_id: arbya.id, approved: true, is_formal: true)
     a9=Synonym.create(name: "سيتنجس", keyword_id: edadat.id, approved: true, is_formal: false)
     a9=Synonym.create(name: "انجاز كتير", keyword_id: engazat.id, approved: true, is_formal: false)
-    b1=Synonym.create(name: "يحمل", keyword_id: loading.id, approved: true, is_formal: true)
-    b2=Synonym.create(name: "لودينج", keyword_id: loading.id, approved: true, is_formal: false)
-    b3=Synonym.create(name: "شاريكوا", keyword_id: share.id, approved: true, is_formal: true)
-    b4=Synonym.create(name: "دوررليا", keyword_id: find.id, approved: true, is_formal: false)
+    b1=Synonym.create(name: "يحمل", keyword_id: loading.id, approved: true, is_formal: nil)
+    b2=Synonym.create(name: "لودينج", keyword_id: loading.id, approved: true, is_formal: nil)
+    b3=Synonym.create(name: "شاريكوا", keyword_id: share.id, approved: true, is_formal: nil)
+    b4=Synonym.create(name: "دوررليا", keyword_id: find.id, approved: true, is_formal: nil)
     b5=Synonym.create(name: "ابحثلي", keyword_id: search.id, approved: true, is_formal: true)
     b5=Synonym.create(name: "سيرشلي", keyword_id: search.id, approved: true, is_formal: false)
